Rewrite image validation code into new patterns
