--- core/src/main/kotlin/org/dreamexposure/discal/core/utils/ImageValidator.kt
@@ -1,61 +1,0 @@
-package org.dreamexposure.discal.core.utils
-
-import reactor.core.publisher.Mono
-import reactor.core.scheduler.Schedulers
-import java.io.FileNotFoundException
-import java.io.IOException
-import java.io.InputStream
-import java.net.MalformedURLException
-import java.net.URI
-import java.time.Duration
-import javax.imageio.ImageIO
-import javax.imageio.ImageReader
-
-object ImageValidator {
-    fun validate(url: String, allowGif: Boolean): Mono<Boolean> {
-        return Mono.fromCallable {
-            val image = ImageIO.read(URI.create(url).toURL())
-            image != null
-        }.subscribeOn(Schedulers.boundedElastic())
-                .onErrorResume(IOException::class.java) {
-                    if (allowGif) validateGif(url)
-                    else Mono.just(false)
-                }.onErrorReturn(MalformedURLException::class.java, false)
-                .onErrorReturn(FileNotFoundException::class.java, false)
-    }
-
-    private fun validateGif(url: String): Mono<Boolean> {
-        return Mono.fromCallable {
-            val connection = URI.create(url).toURL().openConnection()
-            connection.connectTimeout = Duration.ofSeconds(3).toMillis().toInt()
-            connection.readTimeout = Duration.ofSeconds(3).toMillis().toInt()
-
-            connection.getInputStream()
-        }.subscribeOn(Schedulers.boundedElastic())
-                .flatMap(ImageValidator::readGif)
-                .map { it.equals(".gif", true) }
-    }
-
-    private fun readGif(input: InputStream): Mono<String> {
-        return Mono.fromCallable {
-            val stream = ImageIO.createImageInputStream(input)
-            val iterator = ImageIO.getImageReaders(stream)
-            if (!iterator.hasNext())
-                return@fromCallable null
-
-            var reader: ImageReader? = null
-            try {
-                reader = iterator.next()
-
-                reader.setInput(stream, true, true)
-                reader.read(0, reader.defaultReadParam)
-            } catch (_: Exception) {
-
-            } finally {
-                reader?.dispose()
-            }
-            if (reader == null) return@fromCallable "invalid_file_type"
-            return@fromCallable reader.formatName
-        }
-    }
-}
